Consolidate config
