+++ seed/setup/clean.js
@@ -1,21 +1,12 @@
 import S3 from 'aws-sdk/clients/s3.js'
 
-import storageCredentials from '../config.js'
+import awsConfig from '../aws-config.js'
 
-const s3 = new S3({
-  signatureVersion: 'v4',
-  apiVersion: '2006-03-01',
-  endpoint: storageCredentials.endpoint,
-  region: storageCredentials.region,
-  accessKeyId: storageCredentials.accessKeyId,
-  secretAccessKey: storageCredentials.secretAccessKey,
-  s3ForcePathStyle: storageCredentials.forcePathStyle,
-  maxRetries: 3,
-})
+const s3 = new S3(awsConfig);
 
 async function doesBucketExist() {
   try {
-    await s3.headBucket({ Bucket: storageCredentials.bucket }).promise()
+    await s3.headBucket({ Bucket: awsConfig.bucket }).promise()
     return true
   }
   catch (error) {
@@ -27,12 +18,12 @@ async function doesBucketExist() {
 
 async function listAllFiles() {
   if (!await doesBucketExist()) {
-    await s3.createBucket({ Bucket: storageCredentials.bucket }).promise()
+    await s3.createBucket({ Bucket: awsConfig.bucket }).promise()
     return
   }
   return s3
     .listObjectsV2({
-      Bucket: storageCredentials.bucket,
+      Bucket: awsConfig.bucket,
       Prefix: '',
     })
     .promise()
@@ -43,7 +34,7 @@ function deleteFiles(files) {
   if (!files.length)
     return Promise.resolve()
   return s3.deleteObjects({
-    Bucket: storageCredentials.bucket,
+    Bucket: awsConfig.bucket,
     Delete: { Objects: files },
   }).promise()
 }
